Salvando número de ordem do pedido - 100

+++ lib/models/product.dart
@@ -10,7 +10,6 @@ class Product extends ChangeNotifier {
     images = images ?? [];
     sizes = sizes ?? [];
   }
-
   Product.fromDocument(DocumentSnapshot document) {
     id = document.documentID;
     name = document['name'] as String;
@@ -20,20 +19,16 @@ class Product extends ChangeNotifier {
         .map((s) => ItemSize.fromMap(s as Map<String, dynamic>))
         .toList();
   }
-
   final Firestore firestore = Firestore.instance;
   final FirebaseStorage storage = FirebaseStorage.instance;
   DocumentReference get firestoreRef => firestore.document('products/$id');
   StorageReference get storageRef => storage.ref().child('products').child(id);
-
   String id;
   String name;
   String description;
   List<String> images;
   List<ItemSize> sizes;
-
   List<dynamic> newImages;
-
   bool _loading = false;
   bool get loading => _loading;
   set loading(bool value) {
@@ -42,9 +37,7 @@ class Product extends ChangeNotifier {
   }
 
   ItemSize _selectedSize;
-
   ItemSize get selectedSize => _selectedSize;
-
   set selectedSize(ItemSize value) {
     _selectedSize = value;
     notifyListeners();
@@ -89,18 +82,13 @@ class Product extends ChangeNotifier {
       'description': description,
       'sizes': exportSizeList(),
     };
-
     if (id == null) {
       final doc = await firestore.collection('products').add(data);
       id = doc.documentID;
     } else {
       await firestoreRef.updateData(data);
     }
-
-    //images [url1, url2, url3]
-
     final List<String> updateImages = [];
-
     for (final newImage in newImages) {
       if (images.contains(newImage)) {
         updateImages.add(newImage as String);
@@ -114,7 +102,7 @@ class Product extends ChangeNotifier {
     }
 
     for (final image in images) {
-      if (!newImages.contains(image)) {
+      if (!newImages.contains(image) && image.contains('firebase')) {
         try {
           final ref = await storage.getReferenceFromUrl(image);
           await ref.delete();
@@ -123,7 +111,6 @@ class Product extends ChangeNotifier {
         }
       }
     }
-
     await firestoreRef.updateData({'images': updateImages});
     images = updateImages;
     loading = false;
